backend/requirements.txt
fastapi
uvicorn
qdrant-client
torch
torchvision
pillow
numpy
scikit-learn
python-multipart
pytorch-metric-learning
